style: styling the hero banner of the webpage
Used a gradient color on the background image. Added a 'call to action'
div that with arrow icons and text. The user will click this to start
there adventure planning. Changed the title of the hero banner.

Chore: installed html-loader and configured to change the source
of the images for the html file in dist folder