feat: When user searchs for a location for the first time the form and results perform nice animation.
When results are returned the form and header slide up to the top of
the section from the center of the page and the results card fades
in to view. This is achieved using CSS classes transformStartPosition
transformEndPostion, that move the form and header. This classes are
added / removed in formHandler.js, an event listener, listens for the
end of the animation and fires then adds the flex-display class to the
results so it fades in.