feat: Have different animation on the form depending on if its a first search or a subsequent search.
When the user firsts for the first time a loader spinner appears, then
when the results return the form and the title move up and the spinners
disappears and the results card appear. On subsequent searchs the
results disappear and the spinner loader appears then when the results
return the spinner loader disappear and the results appear. This
difference in this animation is the tranlation of the spinner.