feat: When the user clicks add trip on the results card, a card is added to the my trip section.
When the results are returned from the server and display on the results
card, the response object is also saved as a global variable in the
trip js file. When the user clicks the add to trip button on results
card a html template of a results card is created which includes the
parsed results from the response object. This template is converted into
an element and appended to the trip card section of the my trip list.