style and feat: Added animations for a better user experience. Added a add new trip card.
When user click add new trip page for the first time, it is made visible
and the page scroll smoothly to that section. The newly add trip card
fades into view. Also using a delay function the form is reset so that
the result card is hidden and the form and title are centered in the
formSection page.

A simple card that allows the user to add a new trip has been Added
but no functionality assigned to it yet.