feat: Error handling on server and client
Server now has error handling, if there is an issue with any of the
function involved in fetching and parsing the clients post query
to the endpoint './query' then a exception is thrown and the server
responses to the client with a status failure and the an message
detailing the error.

The client handles the failed response by stopping the spinner loader,
reseting the form and an error message over the form section.