refactor: change the way results were handled
When results are returned from the users search the object is now
saved to a global variable in formHandler. formHandler has a exported
function that returns the global variable storing the results.
The trip js file imports this function and uses it to acquire the latest
results when a card is being added.